Handle nodes without coordinates
Treat nodes without coordinates the same as if they are absent, in
construct_multipolygon, construct_multilinestring, and when processing
ways, to avoid producing coordinates of NaN/null.

When a segment of a relation (either a line string, or a polygon ring)
consists entirely of invalid nodes, omit this empty array, to avoid
producing a coordinates element which is an empty array.

Both of these changes will cause such features to be omitted from the
output when they have no coordinates, just as they already are for ways
which reference nodes that are absent, and relations which reference
ways that are absent.